Extração de breadcrumbs de app para classe isolada.
